Refactor .gitignore to include build directory
