+++ src/components/ParticlesComponent.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import Particles, { initParticlesEngine } from "@tsparticles/react";
 import { loadSlim } from "@tsparticles/slim";
 
@@ -13,8 +13,7 @@ const ParticlesComponent = () => {
         });
     }, []);
 
-    const particlesLoaded = (container) => {
-        console.log(container);
+    const particlesLoaded = () => {
     };
 
     const options = useMemo(
