+++ src/context/AuthContext.jsx
@@ -1,0 +1,29 @@
+import React, { createContext, useState, useEffect, useContext } from 'react';
+
+// Create a context for authentication
+export const AuthContext = createContext();
+
+// Custom hook to use the AuthContext
+export const useAuth = () => {
+  return useContext(AuthContext);
+};
+
+export const AuthProvider = ({ children }) => {
+  const [user, setUser] = useState({});
+
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+
+  useEffect(() => {
+    if (Object.keys(user).length > 0) {
+      setIsAuthenticated(true);
+    } else {
+      setIsAuthenticated(false);
+    }
+  }, [user]);
+
+  return (
+    <AuthContext.Provider value={{ user, isAuthenticated, setUser, setIsAuthenticated }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
