+++ src/pages/CheckingOfficerDashboard.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { Button, Modal } from 'antd';
 import { SwapRightOutlined, InfoCircleOutlined, ExclamationCircleOutlined } from '@ant-design/icons';
 import BarcodeScanner from '../components/BarcodeScanner'; // Adjust the path as needed
